backend/requirements.txt
# VacanceAI Backend Dependencies

# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Oracle Database
oracledb>=2.0.0

# JWT Authentication
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# LangChain / LangGraph
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0

# Settings
pydantic[email]>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
email-validator>=2.0.0

# HTTP Client
httpx>=0.28.0

# WebSocket
websockets>=13.0

# Utilities
python-multipart>=0.0.9

# OpenTelemetry
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-otlp-proto-http>=1.20.0
